undefined behavior: signed integer overflow

+++ cups/ppd.c
@@ -2839,7 +2839,7 @@ ppd_hash_option(ppd_option_t *option)	/* I - Option */
 
 
   for (hash = option->keyword[0], k = option->keyword + 1; *k;)
-    hash = 33 * hash + *k++;
+    hash = (int)(33U * (unsigned)hash) + *k++;
 
   return (hash & 511);
 }
